Separate functions to various classes
